Avoid conversions to float
Q and D are stored as doubles in the tree. The computation converts them
to floats. Conversions have a potential for accuracy lose when
calculating delta updates.

+++ src/search/dag_classic/node.cc
@@ -343,7 +343,7 @@ void Node::CancelScoreUpdate(uint32_t multivisit) {
   n_in_flight_.fetch_sub(multivisit, std::memory_order_acq_rel);
 }
 
-void LowNode::FinalizeScoreUpdate(float v, float d, float m,
+void LowNode::FinalizeScoreUpdate(double v, double d, float m,
                                   uint32_t multivisit) {
   assert(edges_);
   // Recompute Q.
@@ -357,7 +357,7 @@ void LowNode::FinalizeScoreUpdate(float v, float d, float m,
   n_ += multivisit;
 }
 
-void LowNode::AdjustForTerminal(float v, float d, float m,
+void LowNode::AdjustForTerminal(double v, double d, float m,
                                 uint32_t multivisit) {
   assert(static_cast<uint32_t>(multivisit) <= n_);
 
@@ -369,7 +369,7 @@ void LowNode::AdjustForTerminal(float v, float d, float m,
   assert(WLDMInvariantsHold());
 }
 
-void Node::FinalizeScoreUpdate(float v, float d, float m, uint32_t multivisit) {
+void Node::FinalizeScoreUpdate(double v, double d, float m, uint32_t multivisit) {
   // Recompute Q.
   wl_ += multivisit * (v - wl_) / (n_ + multivisit);
   d_ += multivisit * (d - d_) / (n_ + multivisit);
@@ -384,7 +384,7 @@ void Node::FinalizeScoreUpdate(float v, float d, float m, uint32_t multivisit) {
   n_in_flight_.fetch_sub(multivisit, std::memory_order_acq_rel);
 }
 
-void Node::AdjustForTerminal(float v, float d, float m, uint32_t multivisit) {
+void Node::AdjustForTerminal(double v, double d, float m, uint32_t multivisit) {
   assert(static_cast<uint32_t>(multivisit) <= n_);
 
   // Recompute Q.
@@ -610,10 +610,10 @@ void Node::SortEdges() const {
   low_node_->SortEdges();
 }
 
-static constexpr float wld_tolerance = 0.000001f;
+static constexpr double wld_tolerance = 0.000001f;
 static constexpr float m_tolerance = 0.000001f;
 
-static bool WLDMInvariantsHold(float wl, float d, float m) {
+static bool WLDMInvariantsHold(double wl, double d, float m) {
   return -(1.0f + wld_tolerance) < wl && wl < (1.0f + wld_tolerance) &&  //
          -(0.0f + wld_tolerance) < d && d < (1.0f + wld_tolerance) &&    //
          -(0.0f + m_tolerance) < m &&                                    //
